requirements.txt
google-adk==1.14.0
langchain-google-vertexai
langchain-community
wikipedia==1.4.0
python-dotenv
sqlalchemy
pg8000
google-api-python-client
google-auth-oauthlib
google-cloud-secret-manager
